move calc to calcthread
API call returns data now
