+++ src/news.py
@@ -6,11 +6,3 @@ class News(object):
         self.datePublished = datePublished
         self.location = location
         self.sentiment = sentiment
-    #def __repr__(self):
-        #return #u'name: ' + self.name#; +\
-               #u'url: ' + utf(self.url) +\
-               #u'description: ' + utf(self.description) +\
-               #u'datePublished: ' + utf(self.datePublished) +\
-               #u'location: ' + (self.location if self.location is not None else u'N/A' )+\
-               #u'sentiment: ' + (self.sentiment if self.location is not None else u'N/A');
-
